day1: part1 and part2
Pretty naive solution but I doubt you can do this without collecting the
numbers, in part 1 due to the sorting requirement, and in part 2 since
you have to keep track of how many times each number appears on the
right.

I can't think of a way to do that without iterating over the input a
gazillion times.